ml_pipeline/requirements-ml.txt
ultralytics>=8.3.0
opencv-python>=4.9.0
torch>=2.2.0
torchvision>=0.17.0
Pillow>=10.0.0
fiftyone>=1.3.0
kagglehub>=0.3.0
tqdm>=4.66.0
pandas>=2.2.0
scikit-learn>=1.4.0
albumentations>=1.4.0
